+++ packages/js-dapi-client/lib/methods/platform/PlatformMethodsFacade.js
@@ -1,9 +1,9 @@
 const broadcastStateTransitionFactory = require('./broadcastStateTransitionFactory');
 const getDataContractFactory = require('./getDataContractFactory');
 const getDocumentsFactory = require('./getDocumentsFactory');
-const getIdentityByFirstPublicKeyFactory = require('./getIdentityByFirstPublicKeyFactory');
 const getIdentityFactory = require('./getIdentityFactory');
-const getIdentityIdByFirstPublicKeyFactory = require('./getIdentityIdByFirstPublicKeyFactory');
+const getIdentityIdsByPublicKeyHashesFactory = require('./getIdentityIdsByPublicKeyHashesFactory');
+const getIdentitiesByPublicKeyHashesFactory = require('./getIdentitiesByPublicKeyHashesFactory');
 
 class PlatformMethodsFacade {
   /**
@@ -13,9 +13,9 @@ class PlatformMethodsFacade {
     this.broadcastStateTransition = broadcastStateTransitionFactory(grpcTransport);
     this.getDataContract = getDataContractFactory(grpcTransport);
     this.getDocuments = getDocumentsFactory(grpcTransport);
-    this.getIdentityByFirstPublicKey = getIdentityByFirstPublicKeyFactory(grpcTransport);
     this.getIdentity = getIdentityFactory(grpcTransport);
-    this.getIdentityIdByFirstPublicKey = getIdentityIdByFirstPublicKeyFactory(grpcTransport);
+    this.getIdentityIdsByPublicKeyHashes = getIdentityIdsByPublicKeyHashesFactory(grpcTransport);
+    this.getIdentitiesByPublicKeyHashes = getIdentitiesByPublicKeyHashesFactory(grpcTransport);
   }
 }