+++ packages/js-dapi-client/lib/methods/platform/getDocumentsFactory.js
@@ -16,7 +16,7 @@ function getDocumentsFactory(grpcTransport) {
    * Fetch Documents from Drive
    *
    * @typedef {getDocuments}
-   * @param {string} contractId - Data Contract ID
+   * @param {Buffer} contractId - Data Contract ID
    * @param {string} type - Document type
    * @param {DAPIClientOptions & getDocumentsOptions} [options]
    * @returns {Promise<Buffer[]>}
@@ -41,6 +41,17 @@ function getDocumentsFactory(grpcTransport) {
     }
 
     const getDocumentsRequest = new GetDocumentsRequest();
+    // need to convert Identifier to pure buffer as google protobuf doesn't support extended buffers
+    // https://github.com/protocolbuffers/protobuf/blob/master/js/binary/utils.js#L1049
+
+    // need to convert objects inherited from Buffer to pure buffer as google protobuf
+    // doesn't support extended buffers
+    // https://github.com/protocolbuffers/protobuf/blob/master/js/binary/utils.js#L1049
+    if (Buffer.isBuffer(contractId)) {
+      // eslint-disable-next-line no-param-reassign
+      contractId = Buffer.from(contractId);
+    }
+
     getDocumentsRequest.setDataContractId(contractId);
     getDocumentsRequest.setDocumentType(type);
     getDocumentsRequest.setWhere(whereSerialized);