+++ packages/js-dapi-client/test/unit/methods/platform/getDataContractFactory.spec.js
@@ -22,7 +22,7 @@ describe('getDataContractFactory', () => {
     dataContractFixture = getDataContractFixture();
 
     response = new GetDataContractResponse();
-    response.setDataContract(dataContractFixture.serialize());
+    response.setDataContract(dataContractFixture.toBuffer());
 
     grpcTransportMock = {
       request: this.sinon.stub().resolves(response),
@@ -42,13 +42,13 @@ describe('getDataContractFactory', () => {
     const request = new GetDataContractRequest();
     request.setId(contractId);
 
-    expect(grpcTransportMock.request).to.be.calledOnceWithExactly(
+    expect(grpcTransportMock.request.getCall(0).args).to.have.deep.members([
       PlatformPromiseClient,
       'getDataContract',
       request,
       options,
-    );
-    expect(result).to.deep.equal(dataContractFixture.serialize());
+    ]);
+    expect(result).to.deep.equal(dataContractFixture.toBuffer());
   });
 
   it('should return null if data contract not found', async () => {
@@ -64,12 +64,12 @@ describe('getDataContractFactory', () => {
     const request = new GetDataContractRequest();
     request.setId(contractId);
 
-    expect(grpcTransportMock.request).to.be.calledOnceWithExactly(
+    expect(grpcTransportMock.request.getCall(0).args).to.have.deep.members([
       PlatformPromiseClient,
       'getDataContract',
       request,
       options,
-    );
+    ]);
     expect(result).to.equal(null);
   });
 
@@ -80,7 +80,7 @@ describe('getDataContractFactory', () => {
     grpcTransportMock.request.throws(error);
 
     const request = new GetDataContractRequest();
-    request.setId(contractId);
+    request.setId(contractId.toBuffer());
 
     try {
       await getDataContract(contractId, options);