+++ packages/js-dapi-client/test/unit/methods/platform/getIdentitiesByPublicKeyHashesFactory.spec.js
@@ -1,0 +1,81 @@
+const {
+  v0: {
+    PlatformPromiseClient,
+    GetIdentitiesByPublicKeyHashesRequest,
+    GetIdentitiesByPublicKeyHashesResponse,
+  },
+} = require('@dashevo/dapi-grpc');
+
+const getIdentityFixture = require('@dashevo/dpp/lib/test/fixtures/getIdentityFixture');
+
+const getIdentitiesByPublicKeyHashesFactory = require(
+  '../../../../lib/methods/platform/getIdentitiesByPublicKeyHashesFactory',
+);
+
+describe('getIdentitiesByPublicKeyHashesFactory', () => {
+  let grpcTransportMock;
+  let getIdentitiesByPublicKeyHashes;
+  let options;
+  let response;
+  let identityFixture;
+  let publicKeyHash;
+
+  beforeEach(function beforeEach() {
+    identityFixture = getIdentityFixture();
+
+    response = new GetIdentitiesByPublicKeyHashesResponse();
+    response.setIdentitiesList(
+      [identityFixture.toBuffer()],
+    );
+
+    publicKeyHash = identityFixture.getPublicKeyById(1).hash();
+
+    grpcTransportMock = {
+      request: this.sinon.stub().resolves(response),
+    };
+
+    options = {
+      timeout: 1000,
+    };
+
+    getIdentitiesByPublicKeyHashes = getIdentitiesByPublicKeyHashesFactory(grpcTransportMock);
+  });
+
+  it('should return public key hashes to identity map', async () => {
+    const result = await getIdentitiesByPublicKeyHashes([publicKeyHash], options);
+
+    const request = new GetIdentitiesByPublicKeyHashesRequest();
+    request.setPublicKeyHashesList([publicKeyHash]);
+
+    expect(grpcTransportMock.request).to.be.calledOnceWithExactly(
+      PlatformPromiseClient,
+      'getIdentitiesByPublicKeyHashes',
+      request,
+      options,
+    );
+    expect(result).to.have.deep.members([identityFixture.toBuffer()]);
+  });
+
+  it('should throw unknown error', async () => {
+    const error = new Error('Unknown found');
+
+    grpcTransportMock.request.throws(error);
+
+    const request = new GetIdentitiesByPublicKeyHashesRequest();
+    request.setPublicKeyHashesList([publicKeyHash]);
+
+    try {
+      await getIdentitiesByPublicKeyHashes([publicKeyHash], options);
+
+      expect.fail('should throw unknown error');
+    } catch (e) {
+      expect(e).to.deep.equal(error);
+      expect(grpcTransportMock.request).to.be.calledOnceWithExactly(
+        PlatformPromiseClient,
+        'getIdentitiesByPublicKeyHashes',
+        request,
+        options,
+      );
+    }
+  });
+});